refactor(docz): fix error with docz and finalize
