+++ docs/src/pages/checkbox.mdx
@@ -5,8 +5,7 @@ menu: Components
 ---
 
 import { Playground, Props } from "docz"
-
-import Checkbox from './'
+import Checkbox from '../../packages/components/Checkbox'
 
 # Checkbox
 
@@ -21,6 +20,10 @@ The HTML DOM structure is
   <span />       // Holds the label (override - label)
 </label>
 ```
+
+## Props
+
+<Props of={Checkbox}/>
 
 ## Default Checkbox
 <Playground>
