--- docs/src/pages/checkbox/index.js
@@ -1,6 +1,0 @@
-import { Checkbox as BaseCheckbox } from '../../../../src/components/Checkbox';
-import { withTheme } from '../../components/withTheme';
-
-const Checkbox = withTheme(BaseCheckbox);
-
-export default Checkbox;
