+++ docs/src/pages/text.mdx
@@ -4,9 +4,8 @@ route: /components/text
 menu: Components
 ---
 
-import { Playground } from "docz"
-
-import Text from './'
+import { Props, Playground } from "docz"
+import Text from '../../packages/components/Text'
 
 # Text
 
@@ -18,6 +17,10 @@ The HTML DOM structure is
 // Override the base styles using the `override` prop
 <Text override={}/>
 ```
+
+## Props
+
+<Props of={Text}/>
 
 ## Basic Unstyled Text
